requirements.txt
Flask==3.0.0
pymongo==4.6.0
flask-pymongo==2.3.0
paho-mqtt==1.6.1
openai==1.3.0
python-dotenv==1.0.0
werkzeug==3.0.1
bcrypt==4.1.1

